# Conflicts:
#	src/main/java/net/atlas/combatify/mixin/LivingEntityMixin.java
#	src/main/java/net/atlas/combatify/util/blocking/BlockingType.java
#	src/main/java/net/atlas/combatify/util/blocking/CurrentShieldBlockingType.java
#	src/main/java/net/atlas/combatify/util/blocking/EmptyBlockingType.java
#	src/main/java/net/atlas/combatify/util/blocking/NewShieldBlockingType.java
#	src/main/java/net/atlas/combatify/util/blocking/OldSwordBlockingType.java
#	src/main/java/net/atlas/combatify/util/blocking/PercentageBlockingType.java
#	src/main/java/net/atlas/combatify/util/blocking/ShieldBlockingType.java
#	src/main/java/net/atlas/combatify/util/blocking/SwordBlockingType.java
#	src/main/java/net/atlas/combatify/util/blocking/TestBlockingType.java
#	src/main/java/net/atlas/combatify/util/blocking/condition/ItemMatches.java
#	src/main/java/net/atlas/combatify/util/blocking/effect/HurtEntity.java

+++ src/main/java/net/atlas/combatify/extensions/Tier.java
@@ -22,7 +22,7 @@ public interface Tier {
 		throw new IllegalStateException("Extension has not been applied");
 	}
 
-	default float combatify$blockingLevel() {
+	default int combatify$blockingLevel() {
 		throw new IllegalStateException("Extension has not been applied");
 	}
 }
